requirements.txt
fastapi
uvicorn
python-dotenv
requests
groq
transformers
torch
diffusers
Pillow